refactor: join static and dynamic configs to reuse `pollQueries()`
