+++ lib/graphql/types.js
@@ -140,8 +140,12 @@ type DynamicConfig {
   restartServices: Boolean!
 }
 
+type Configs {
+  static: StaticConfig
+  dynamic: DynamicConfig!
+}
+
 type Query {
-  staticConfig(currentConfigVersion: Int): StaticConfig
-  dynamicConfig: DynamicConfig!
+  configs(currentConfigVersion: Int): Configs!
 }
 `
